requirements.txt
pyyaml
BeautifulSoup